adding files

+++ src/main/java/com/klef/project/service/ApplicationService.java
@@ -7,9 +7,9 @@ import java.util.Optional;
 public interface ApplicationService {
     Application saveApplication(Application application);
     List<Application> getAllApplications();
-    Optional<Application> getApplicationById(Long id);
-    void deleteApplication(Long id);
-    public Application updateStatus(Long applicationId, String newStatus) ;
-    boolean isApplicationExist(String email, Long jobId);  // Add this method to check for existing applications
+    Optional<Application> getApplicationById(Integer id);
+    void deleteApplication(Integer id);
+    public Application updateStatus(Integer applicationId, String newStatus) ;
+    boolean isApplicationExist(String email, Integer jobId);  // Add this method to check for existing applications
     public List<Application> getApplicationsByEmail(String email);
 }
